fastapi-service/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
PyMuPDF==1.23.7
python-docx==1.1.0
google-generativeai==0.3.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
aiofiles==23.2.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
